+++ packages/create-svelte/bin.js
@@ -40,46 +40,48 @@ async function main() {
 		mkdirp(cwd);
 	}
 
-	const options = /** @type {import('./types/internal').Options} */ (await prompts([
-		{
-			type: 'select',
-			name: 'template',
-			message: 'Which Svelte app template?',
-			choices: fs.readdirSync(dist('templates')).map((dir) => {
-				const meta_file = dist(`templates/${dir}/meta.json`);
-				const meta = JSON.parse(fs.readFileSync(meta_file, 'utf8'));
+	const options = /** @type {import('./types/internal').Options} */ (
+		await prompts([
+			{
+				type: 'select',
+				name: 'template',
+				message: 'Which Svelte app template?',
+				choices: fs.readdirSync(dist('templates')).map((dir) => {
+					const meta_file = dist(`templates/${dir}/meta.json`);
+					const meta = JSON.parse(fs.readFileSync(meta_file, 'utf8'));
 
-				return {
-					title: meta.description,
-					value: dir
-				};
-			})
-		},
-		{
-			type: 'toggle',
-			name: 'typescript',
-			message: 'Use TypeScript?',
-			initial: false,
-			active: 'Yes',
-			inactive: 'No'
-		},
-		{
-			type: 'toggle',
-			name: 'eslint',
-			message: 'Add ESLint for code linting?',
-			initial: false,
-			active: 'Yes',
-			inactive: 'No'
-		},
-		{
-			type: 'toggle',
-			name: 'prettier',
-			message: 'Add Prettier for code formatting?',
-			initial: false,
-			active: 'Yes',
-			inactive: 'No'
-		}
-	]));
+					return {
+						title: meta.description,
+						value: dir
+					};
+				})
+			},
+			{
+				type: 'toggle',
+				name: 'typescript',
+				message: 'Use TypeScript?',
+				initial: false,
+				active: 'Yes',
+				inactive: 'No'
+			},
+			{
+				type: 'toggle',
+				name: 'eslint',
+				message: 'Add ESLint for code linting?',
+				initial: false,
+				active: 'Yes',
+				inactive: 'No'
+			},
+			{
+				type: 'toggle',
+				name: 'prettier',
+				message: 'Add Prettier for code formatting?',
+				initial: false,
+				active: 'Yes',
+				inactive: 'No'
+			}
+		])
+	);
 
 	const name = path.basename(path.resolve(cwd));
 
@@ -157,9 +159,9 @@ function write_template_files(template, typescript, name, cwd) {
 	copy(`${dir}/package.json`, `${cwd}/package.json`);
 
 	const manifest = `${dir}/files.${typescript ? 'ts' : 'js'}.json`;
-	const files = /** @type {import('./types/internal').File[]} */ (JSON.parse(
-		fs.readFileSync(manifest, 'utf-8')
-	));
+	const files = /** @type {import('./types/internal').File[]} */ (
+		JSON.parse(fs.readFileSync(manifest, 'utf-8'))
+	);
 
 	files.forEach((file) => {
 		const dest = path.join(cwd, file.name);
@@ -177,9 +179,9 @@ function write_template_files(template, typescript, name, cwd) {
  */
 function write_common_files(cwd, options, name) {
 	const shared = dist('shared.json');
-	const { files } = /** @type {import('./types/internal').Common} */ (JSON.parse(
-		fs.readFileSync(shared, 'utf-8')
-	));
+	const { files } = /** @type {import('./types/internal').Common} */ (
+		JSON.parse(fs.readFileSync(shared, 'utf-8'))
+	);
 
 	const pkg_file = path.join(cwd, 'package.json');
 	const pkg = /** @type {any} */ (JSON.parse(fs.readFileSync(pkg_file, 'utf-8')));