+++ packages/kit/src/core/dev/index.js
@@ -18,7 +18,6 @@ import { print_config_conflicts } from '../config/index.js';
 import { create_app } from '../create_app/index.js';
 import create_manifest_data from '../create_manifest_data/index.js';
 import { getRawBody } from '../node/index.js';
-import { get_server } from '../server/index.js';
 import { SVELTE_KIT, SVELTE_KIT_ASSETS } from '../constants.js';
 import { copy_assets, resolve_entry } from '../utils.js';
 import { coalesce_to_error } from '../../utils/error.js';
@@ -107,13 +106,16 @@ class Watcher extends EventEmitter {
 			}
 		};
 
-		/** @type {(req: import("http").IncomingMessage, res: import("http").ServerResponse) => void} */
-		let handler = (req, res) => {};
-
-		this.server = await get_server(this.https, vite_config, (req, res) => handler(req, res));
-
 		// don't warn on overriding defaults
 		const [modified_vite_config] = deep_merge(default_config, vite_config);
+
+		const kit_plugin = await create_plugin(this.config, this.dir, this.cwd, () => {
+			if (!this.manifest) {
+				throw new Error('Manifest is not available');
+			}
+
+			return this.manifest;
+		});
 
 		/** @type {[any, string[]]} */
 		const [merged_config, conflicts] = deep_merge(modified_vite_config, {
@@ -139,14 +141,13 @@ class Watcher extends EventEmitter {
 					compilerOptions: {
 						hydratable: !!this.config.kit.hydrate
 					}
-				})
+				}),
+				kit_plugin
 			],
 			publicDir: this.config.kit.files.assets,
 			server: {
-				middlewareMode: true,
-				hmr: {
-					...(this.https ? { server: this.server, port: this.port } : {})
-				}
+				host: this.host,
+				https: this.https
 			},
 			base: this.config.kit.paths.assets.startsWith('/') ? `${this.config.kit.paths.assets}/` : '/'
 		});
@@ -154,18 +155,8 @@ class Watcher extends EventEmitter {
 		print_config_conflicts(conflicts, 'kit.vite.');
 
 		this.vite = await vite.createServer(merged_config);
-
-		const get_manifest = () => {
-			if (!this.manifest) {
-				throw new Error('Manifest is not available');
-			}
-
-			return this.manifest;
-		};
-
-		handler = await create_handler(this.vite, this.config, this.dir, this.cwd, get_manifest);
-
-		this.server.listen(this.port, this.host || '0.0.0.0');
+		remove_html_middlewares(this.vite.middlewares);
+		await this.vite.listen(this.port);
 	}
 
 	update() {
@@ -212,7 +203,7 @@ class Watcher extends EventEmitter {
 	}
 
 	close() {
-		if (!this.vite || !this.server || !this.cheapwatch) {
+		if (!this.vite || !this.cheapwatch) {
 			throw new Error('Cannot close server before it is initialized');
 		}
 
@@ -220,7 +211,6 @@ class Watcher extends EventEmitter {
 		this.closed = true;
 
 		this.vite.close();
-		this.server.close();
 		this.cheapwatch.close();
 	}
 }
@@ -231,15 +221,31 @@ function get_params(array) {
 	// src/routes/[x]/[y]/[z]/svelte, create a function
 	// that turns a RegExpExecArray into ({ x, y, z })
 
+	// input has already been decoded by decodeURI
+	// now handle the rest that decodeURIComponent would do
+	const d = /** @param {string} s */ (s) =>
+		s
+			.replace(/%23/g, '#')
+			.replace(/%3[Bb]/g, ';')
+			.replace(/%2[Cc]/g, ',')
+			.replace(/%2[Ff]/g, '/')
+			.replace(/%3[Ff]/g, '?')
+			.replace(/%3[Aa]/g, ':')
+			.replace(/%40/g, '@')
+			.replace(/%26/g, '&')
+			.replace(/%3[Dd]/g, '=')
+			.replace(/%2[Bb]/g, '+')
+			.replace(/%24/g, '$');
+
 	/** @param {RegExpExecArray} match */
 	const fn = (match) => {
 		/** @type {Record<string, string>} */
 		const params = {};
 		array.forEach((key, i) => {
 			if (key.startsWith('...')) {
-				params[key.slice(3)] = decodeURIComponent(match[i + 1] || '');
+				params[key.slice(3)] = d(match[i + 1] || '');
 			} else {
-				params[key] = decodeURIComponent(match[i + 1]);
+				params[key] = d(match[i + 1]);
 			}
 		});
 		return params;
@@ -249,13 +255,12 @@ function get_params(array) {
 }
 
 /**
- * @param {vite.ViteDevServer} vite
  * @param {import('types/config').ValidatedConfig} config
  * @param {string} dir
  * @param {string} cwd
  * @param {() => import('types/internal').SSRManifest} get_manifest
  */
-async function create_handler(vite, config, dir, cwd, get_manifest) {
+async function create_plugin(config, dir, cwd, get_manifest) {
 	/**
 	 * @type {amp_validator.Validator?}
 	 */
@@ -275,11 +280,14 @@ async function create_handler(vite, config, dir, cwd, get_manifest) {
 	};
 
 	/**
-	 * @param {import('http').IncomingMessage} req
-	 * @param {import('http').ServerResponse} res
+	 * @param {vite.ViteDevServer} vite
 	 */
-	return (req, res) => {
-		vite.middlewares(req, res, async () => {
+	function create_kit_middleware(vite) {
+		/**
+		 * Use a named function for debugging
+		 * @type {import('connect').NextHandleFunction}
+		 */
+		return async function svelteKitMiddleware(req, res) {
 			try {
 				if (!req.url || !req.method) throw new Error('Incomplete request');
 				if (req.url === '/favicon.ico') return not_found(res);
@@ -340,8 +348,9 @@ async function create_handler(vite, config, dir, cwd, get_manifest) {
 					return res.end(err.reason || 'Invalid request body');
 				}
 
-				const host = /** @type {string} */ (config.kit.host ||
-					req.headers[config.kit.hostHeader || 'host']);
+				const host = /** @type {string} */ (
+					config.kit.host || req.headers[config.kit.hostHeader || 'host']
+				);
 
 				const rendered = await respond(
 					{
@@ -492,7 +501,19 @@ async function create_handler(vite, config, dir, cwd, get_manifest) {
 				res.statusCode = 500;
 				res.end(error.stack);
 			}
-		});
+		};
+	}
+
+	return {
+		name: 'vite-plugin-svelte-kit',
+		/**
+		 * @param {import('vite').ViteDevServer} vite
+		 */
+		configureServer(vite) {
+			return () => {
+				vite.middlewares.use(create_kit_middleware(vite));
+			};
+		}
 	};
 }
 
@@ -500,4 +521,21 @@ async function create_handler(vite, config, dir, cwd, get_manifest) {
 function not_found(res) {
 	res.statusCode = 404;
 	res.end('Not found');
+}
+
+/**
+ * @param {import('connect').Server} server
+ */
+function remove_html_middlewares(server) {
+	const html_middlewares = [
+		'viteIndexHtmlMiddleware',
+		'vite404Middleware',
+		'viteSpaFallbackMiddleware'
+	];
+	for (let i = server.stack.length - 1; i > 0; i--) {
+		// @ts-expect-error using internals until https://github.com/vitejs/vite/pull/4640 is merged
+		if (html_middlewares.includes(server.stack[i].handle.name)) {
+			server.stack.splice(i, 1);
+		}
+	}
 }