+++ packages/kit/src/packaging/test/fixtures/exports-replace/svelte.config.js
@@ -2,9 +2,7 @@
 const config = {
 	kit: {
 		package: {
-			exports: {
-				exclude: ['**']
-			}
+			exports: () => false
 		}
 	}
 };