+++ packages/kit/src/packaging/test/fixtures/resolve-alias/expected/baz.d.ts
@@ -1,0 +1,4 @@
+export interface Baz {
+	baz: string;
+}
+export declare const baz: Baz;