+++ packages/kit/src/packaging/test/fixtures/resolve-alias/svelte.config.cjs
@@ -1,0 +1,5 @@
+const preprocess = require('svelte-preprocess');
+
+module.exports = {
+	preprocess: preprocess()
+};