+++ packages/kit/src/packaging/test/fixtures/typescript/expected/Test2.svelte
@@ -1,3 +1,3 @@
-<script lang="ts">
+<script>
     export let foo;
 </script>