+++ packages/kit/src/runtime/server/page/respond.js
@@ -73,7 +73,7 @@ export async function respond(opts) {
 	let set_cookie_headers = [];
 
 	ssr: if (page_config.ssr) {
-		let context = {};
+		let stuff = {};
 
 		for (let i = 0; i < nodes.length; i += 1) {
 			const node = nodes[i];
@@ -86,7 +86,7 @@ export async function respond(opts) {
 					loaded = await load_node({
 						...opts,
 						node,
-						context,
+						stuff,
 						prerender_enabled: is_prerender_enabled(options, node, state),
 						is_leaf: i === nodes.length - 1,
 						is_error: false
@@ -138,16 +138,18 @@ export async function respond(opts) {
 
 							try {
 								// there's no fallthough on an error page, so we know it's not undefined
-								const error_loaded = /** @type {import('./types').Loaded} */ (await load_node({
-									...opts,
-									node: error_node,
-									context: node_loaded.context,
-									prerender_enabled: is_prerender_enabled(options, error_node, state),
-									is_leaf: false,
-									is_error: true,
-									status,
-									error
-								}));
+								const error_loaded = /** @type {import('./types').Loaded} */ (
+									await load_node({
+										...opts,
+										node: error_node,
+										stuff: node_loaded.stuff,
+										prerender_enabled: is_prerender_enabled(options, error_node, state),
+										is_leaf: false,
+										is_error: true,
+										status,
+										error
+									})
+								);
 
 								if (error_loaded.loaded.error) {
 									continue;
@@ -183,11 +185,10 @@ export async function respond(opts) {
 				}
 			}
 
-			if (loaded && loaded.loaded.context) {
-				// TODO come up with better names for stuff
-				context = {
-					...context,
-					...loaded.loaded.context
+			if (loaded && loaded.loaded.stuff) {
+				stuff = {
+					...stuff,
+					...loaded.loaded.stuff
 				};
 			}
 		}