+++ packages/kit/test/apps/basics/src/routes/load/_tests.js
@@ -71,7 +71,7 @@ export default function (test, is_dev) {
 		assert.equal(await page.textContent('h1'), 'static file');
 	});
 
-	test('context is inherited', '/load/context/a/b/c', async ({ page, js, app }) => {
+	test('stuff is inherited', '/load/stuff/a/b/c', async ({ page, js, app }) => {
 		assert.equal(await page.textContent('h1'), 'message: original + new');
 		assert.equal(
 			await page.textContent('pre'),
@@ -83,7 +83,7 @@ export default function (test, is_dev) {
 		);
 
 		if (js) {
-			await app.goto('/load/context/d/e/f');
+			await app.goto('/load/stuff/d/e/f');
 
 			assert.equal(await page.textContent('h1'), 'message: original + new');
 			assert.equal(